pkgs/litedram: temporarily disable checks b/c of BIOS overrun
Checks are currently disabled for the LiteDRAM package, as there is a
BIOS size overrun for one tested board. The tests should be reenabled
on the next test. One may also experiment with custom compiler /
glibc (newlib, etc.)  options and see whether they have any effect or
are used by the BIOS.

Signed-off-by: Leon Schuermann <leon@is.currently.online>