pkgs/litedram: reenable checks with the failing check deselected
Furthermore, the failing check is asserted to fail, which will serve
as a reminder to remove the exception once the genesys2 test starts to
work again.

Also changes the litevideo and valentyusb-hw_cdc_eptri packages to use
an unchecked LiteX derivation as a build input to avoid producing
infinite recursion across dependencies.

Signed-off-by: Leon Schuermann <leon@is.currently.online>