maintenance/update_packages: allow specifying TOML path
To increase the utility of this maintenance script in other
repositories, allow specifiying the path to the TOML file containing
the package meta data as a command line parameter.

Furthermore, this changes this script such that it overwrites the
source files. This seems like the expected behavior for such an update
script and users are expected to employ some version control system in
their repositories which can recover the old version of the file in
case of unintended changes.

Signed-off-by: Leon Schuermann <leon@is.currently.online>