Add a maintenance script to render a requirements.txt file
This adds a Python script to render a requirements.txt for pip3 to
read in. This can be used to install most of the LiteX and related
Python packages.

This will try to read a TOML file (defaulting to
./pkgs/litex_packages.toml), if necessary fetch required intermediate
information, and finally build a pip3-compatible requirements.txt
file.

This script is mostly useful for downstream repositories which would
like the Nix support and test suite provided by nix-litex, but offer
users the option to get the packages without using the Nix package
manager.

Signed-off-by: Leon Schuermannn <leon@is.currently.online>