vivado: revert to version 2020.1
This reverts Vivado back to a tested derivation for version 2020.1.
Unfortunately, the current derivation for 2020.2 -- while building
successfully -- hangs at startup. While this should be fixed and Vivado
2020.2 should be supported, it is better to have a confirmedworking
version 2020.1 in the repository.

Ideally, many past versions of Vivado should be supported, given that
there may be behavior specific to one version which is relied on, and
compatibility between different components of Vivado is very limited
(e.g. the hardware server needs to have a matching version, etc.). Thus
this renames the derivation and builder script accordingly.

Signed-off-by: Leon Schuermann <leon@is.currently.online>
Reviewed-by: Las Safin <me@las.rs>