Update LiteX packages to the latest revisions
This is not an entirely trivial update, given that the packages' test
infrastructure has advanced significantly and thus required some
changes in terms of Nix packaging.

LiteX has added integration tests for various softcore CPUs. Given
that this repository only contains definitions for VexRiscv and SERV,
and the VexRiscv-SMP CPU requires sbt with an Internet connection to
build, this limits the tested CPUs to the supported ones.

The Genesys2 test of LiteDRAM works again, the BIOS no longer overruns
the allocated ROM size. Thus it is enabled again.

Furthermore, LiteX now depends on picolibc for BIOS
compilation. picolibc itself required a few patches to remove
references to the non-existent /bin/bash interpreter.

This commit uses builtins.fetchGit with the `submodules = true;`
parameter, which is only supported in Nix >= v2.4. As such, this
change bumps the minimum required Nix revision to v2.4.

Signed-off-by: Leon Schuermann <leon@is.currently.online>
Reviewed-by: Las Safin <me@las.rs>