Switch from fetchFromGitHub to builtins.fetchGit
In preparation of derivations requiring a git fetcher which supports
submodules as well, this switches all derivations to use the builtin
fetchGit fetcher instead. An additional benefit of this change is that
it does not require additional recursive hashes of the source. This
makes package updates easier and does not require any git fetching
logic in the maintenance scripts any longer.

To allow users to use their own package sources without extensive
package overrides, a follow-up change to this should generate the src
attributes entirely on the TOML specification.

Signed-off-by: Leon Schuermann <leon@is.currently.online>
Reviewed-by: Las Safin <me@las.rs>