Avoid Python conflict checking in composite (final) derivation
This unbreaks builds of nix-litex packages on recent versions of
nixpkgs, where the pythonCatchConflictsPhase seems to prevent
built. An error such as the one below appears:

    this derivation will be built:
      /nix/store/9hnmqqjvzqxwz871fa2hq9bxy6d54say-python3.9-liteeth-b0e72431234cbbf42a0f74f1a5f2bff145dfc043.drv
    building '/nix/store/9hnmqqjvzqxwz871fa2hq9bxy6d54say-python3.9-liteeth-b0e72431234cbbf42a0f74f1a5f2bff145dfc043.drv'
    Sourcing python-remove-tests-dir-hook
    Sourcing python-catch-conflicts-hook.sh
    Sourcing python-remove-bin-bytecode-hook.sh
    Sourcing setuptools-build-hook
    Using setuptoolsShellHook
    Sourcing pip-install-hook
    Sourcing python-imports-check-hook.sh
    Using pythonImportsCheckPhase
    Sourcing python-namespaces-hook
    unpacking sources
    patching sources
    configuring
    building
    installing
    post-installation fixup
    pythonCatchConflictsPhase
    Found duplicated packages in closure for dependency 'litex':
      litex 0.0.0 (/nix/store/7dwfw7a7rzqnbg52z2w2mrzr01bw422a-python3.9-litex-test-1d20bbcd01784158088e1eb85e219309cfce0248/lib/python3.9/site-packages)
      litex 0.0.0 (/nix/store/zmnzqkjh0lzhhx0g7hwk6b5jh3jgnf4l-python3.9-litex-pkg-1d20bbcd01784158088e1eb85e219309cfce0248/lib/python3.9/site-packages)

    Package duplicates found in closure, see above. Usually this
    happens if two packages depend on different version of the same
    dependency.

This check usually makes sense to ensure that all inputs to a
derivation create an unambigous set of Python packages. However,
within the architecture of nix-litex we deliberately create a
derivation which combines an untested (-pkg) package with its tests
from a separate derivation, as a native build input. The new
derivation just reexposes the untested (-pkg) result, but requires
tests to be run, all without worrying about cyclic dependencies.
Nonetheless, during the build of the final composite derivation, it
has both the results of the untested (-pkg) derivation, as well as the
tested (-test) in its environment. Because this is intended, we
disable this check here. It remains enabled in the untested (-pkg) and
tested (-test) derivations.

Signed-off-by: Leon Schuermann <leon@is.currently.online>