shell.nix: re-introduce Tock Secure IMC VexRiscv variant
This also serves as a demonstration of the way to override the
pythondata-cpu-vexriscv package to automatically generate
something different. This requires that the litex package set
be set up as an overlay in the example shell.nix so that the
internal uses of pythondata-cpu-vexriscv are overriden too and we
don't end up with two different copies.

Signed-off-by: Thomas Watson <twatson52@icloud.com>