pkgs/default.nix: pin a nixpkgs revision for sbt-mkDerivation
This pins a Nixpkgs revision to source the `sbt` package and JRE from,
in order to build softcores written in Scala DSLs (e.g. VexRiscv). We
use `sbt-mkDerivation`, which produces a fixed output hash derivation
of all package dependencies before building the actual package.
However, this intermediate dependency derivation seem to be dependent
on the used Nixpkgs from which sbt, the JRE and all other packages for
the sbt build is taken. Hence, we pin a version here (preferably from
the current Nixpkgs release). It can be overridden by passing the
`sbtNixpkgs` argument.

Signed-off-by: Leon Schuermann <leon@is.currently.online>