Update sbt-derivation to revision of 2022-08-30
This patch updates to the latest version of sbt-derivation which
addresses several shortcomings identified and worked around previously.

The function is now designed to be used alone rather than being an
attribute of sbt. Its new name is mkSbtDerivation.

The location of the sbt dependencies directory no longer needs to be
managed manually by the derivation to account for the unconventional
structure of the VexRiscv repositories as it is now stored in a hidden
global folder and managed entirely by mkSbtDerivation.

The reproducibility issues created by compressing the dependencies using
zstd have been solved by using new configuration options to specify that
they are copied directly. Compression did not save more than 10% anyway.
The hashes have been verified to reproduce all the way back to NixOS
19.03, before which nix-litex fails to evaluate. However, the logic that
separately pins nixpkgs for mkSbtDerivation remains just in case.

Finally, a couple tweaks have been made to pythondata-cpu-vexriscv_smp
to improve build performance and reduce store size.

Signed-off-by: Thomas Watson <twatson52@icloud.com>
Reviewed-by: Leon Schuermann <leon@is.currently.online>