Set `format = "other"` for final derivations to unbreak distPhase
Along with a recent nixpkgs update, the buildPythonPackage function
has been updated to check that there is a ./dist directory in the
built derivations when using setuptools. This thus breaks when
building the final derivations (combining `-unchecked` and `-test`
packages), which simply copy the `-unchecked` package contents to
$out:

    Executing pythonOutputDistPhase
    The build contains no ./dist directory.
    If this project is not setuptools-based, pass

      format = "other";

    to buildPythonApplication/buildPythonPackage or another
    appropriate value as described here:

      https://nixos.org/manual/nixpkgs/stable/#buildpythonpackage-function

Setting the format to "other" as recommended fixes this issue. Given
that the package is already built in the `-unchecked` derivation, this
should avoid most buildPythonPackage magic and simply produce a
package with the pre-built output.

Signed-off-by: Leon Schuermann <leon@is.currently.online>