Add Hydra jobset specification (release.nix)
To support continuous integration building of the nix-litex package
set against various upstream Nixpkgs versions, and to prepare for
integrating an automated update bot with the repository, this adds a
Hydra jobset specification in release.nix.

Currently, this specification is rather minimalistic. It takes in a
reference nixpkgs source to import (provided by Hydra as part of the
jobset definition) and builds all output packages.

Previously, the `default.nix` packages attribute further contained the
`mkSbtDerivation` expression, which the Hydra evaluator does not
like. Also, it seems weird to have the package set be exported both as
the top-level attribute set, and as an inner `packages`
attribute. Hence this further changes the resulting attribute set to
only expose the package set through its respective attribute, and
exposes `mkSbtDerivation` in a separate attribute (given its useful
for downstream users).

Signed-off-by: Leon Schuermann <leon@is.currently.online>