Switch from checkInputs to nativeCheckInputs on >= NixOS 23.05
This unbreaks package tests on the current nixpkgs-unstable branch, as
`buildPythonPackage` has been adjusted to rename `checkInputs` to
`nativeCheckInputs`, breaking existing checkInputs specifications
because of `strictDeps = 1;`. For more information, check out the
corresponding PR at [1]. This implements the proposed workaround[2]
for supporting both the current nixos-unstable and the last release
branch (NixOS 22.11).

[1]: https://github.com/NixOS/nixpkgs/pull/206742
[2]: https://github.com/NixOS/nixpkgs/pull/206742#issuecomment-1417674430

Signed-off-by: Leon Schuermann <leon@is.currently.online>